requirements.txt
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
line-bot-sdk==1.17
pandas==0.25.0
psycopg2==2.8.3
pycparser==2.19
Flask-Login==0.4.1
Pillow==7.2.0
BeautifulSoup4==4.9.1
gunicorn==20.0.4
